refactor<View>: inputView, outputView를 View폴더 안으로
