+++ src/main/java/racingcar/View/InputView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.View;
 
 import java.util.*;
 import java.util.stream.Collectors;
